Actually dealing with image exports in no-export mode.
UI images will be properly exported to the `ui/` directory and their relative URLs will be expanded.
All other image URLs will be nulled out in the content, and won't be added to import. As a side-effect, if they're already present in the theme, they'll be removed.

To do:
Image UI imports has a bunch of code duplication going on. This really needs to be refactored to something saner.